requirements.txt
Django>=2.2.4
PyJWT>=1.7.1
djangorestframework>=3.10.2
django-cors-headers>=3.0.2
Pillow>=6.1.0
